% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{criterionquantiles}
\alias{criterionquantiles}
\title{Calculate criterion quantiles}
\usage{
criterionquantiles(
  y,
  X,
  level,
  estimator_grid,
  nbr_bootstraps = 1000L,
  w = 10L,
  bs_type = "multiplier"
)
}
\arguments{
\item{y}{Response variable vector.}

\item{X}{Covariates matrix.}

\item{level}{Vector of levels.}

\item{estimator_grid}{Lasso estimator grid matrix.}

\item{nbr_bootstraps}{Number of bootstrap simulations. Defaults to 1000.}

\item{w}{Number of blocks. If bs_type = "multiplier" then w is not applicable. If bs_type = "stationary", then w is the mean of number of blocks used for each bootstrap sample. Defaults to 10.}

\item{bs_type}{Type of bootstrap technique to be used. The following methods are available for use: "multiplier", "stationary_block" and "moving_block". Any specified character length is sufficient. Defaults to "multiplier".}
}
\value{
criterion_quantiles Returns a matrix of quantile at each level with each row being the different values from the grid of lambdas.
}
\description{
'criterionquantiles' lets you calculate the criterion quantiles using a bootstrap.
}
\examples{
lasso <- lasso(Y = Y,
               X = X,
               out = 1,
               numlam = 100,
               maxIter = 1e10,
               opTol = 1e-10,
               standardize = 0)

#Remove equidistant lambda row
BETA <- lasso \%>\% head(-1)

criterion_q <- criterionquantiles(Y = Y,
                                  X = X,
                                  level = c(0.90, 0.95, 0.99),
                                  estimator_grid = BETA,
                                  nbr_bootstraps = 10000,
                                  w = 20,
                                  bs_type = "det")
}
