% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lasso}
\alias{lasso}
\title{Compute lasso}
\usage{
lasso(X, Y, out, numlam, maxIter, optTol, standardize)
}
\arguments{
\item{X}{Covariates matrix.}

\item{Y}{Response variable vector.}

\item{out}{1: Outputs betas. 2: Outputs residuals.}

\item{numlam}{Number of lambdas in equidistant lambda grid.}

\item{maxIter}{Maximum number of iterations.}

\item{optTol}{The optimization tolerance as an indicatin for convergence.}

\item{standardize}{Whether or not data should be standardized.}
}
\value{
lasso Returns lasso estimates for each lambda.
}
\description{
'lasso' lets you compute the lasso estimator using the coordinate descent algorithm.
}
\examples{
BETA <- lasso(X = X, Y = Y, out = 1)

}
