% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{lb_folds_custom}
\alias{lb_folds_custom}
\title{Custom fold generator}
\usage{
lb_folds_custom(fn)
}
\arguments{
\item{fn}{A function with signature \verb{function(data) -> integer vector} of
fold IDs the same length as \code{nrow(data)}, with values in \code{1:k} for some
\code{k}. The function is responsible for any re-randomisation across CV
repetitions (typically by not closing over a fixed state).}
}
\value{
The input \code{fn}, validated and classed as an \code{lb_fold_generator}.
}
\description{
Custom fold generator
}
