% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{lb_folds_kfold}
\alias{lb_folds_kfold}
\title{Standard k-fold cross-validation fold generator}
\usage{
lb_folds_kfold(k = 10)
}
\arguments{
\item{k}{Number of folds. Default \code{10}.}
}
\value{
A fold-generator closure \verb{function(data) -> integer vector} of fold
IDs the same length as \code{nrow(data)}.
}
\description{
Standard k-fold cross-validation fold generator
}
