% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lb_plot_complexity}
\alias{lb_plot_complexity}
\title{Plot model complexity (number of selected terms) across bootstrap iterations}
\usage{
lb_plot_complexity(x, ...)
}
\arguments{
\item{x}{An \code{lb_boot} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot model complexity (number of selected terms) across bootstrap iterations
}
