% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantileestimated}
\alias{quantileestimated}
\title{Calculate the estimated quantiles of the effective noise}
\usage{
quantileestimated(Y, X, level, lambda_grid, criterion_quantiles)
}
\arguments{
\item{Y}{Response variable vector.}

\item{X}{Covariates matrix.}

\item{level}{Vector of levels.}

\item{criterion_quantiles}{The criterion quantiles computed using the 'criterionquantiles' function.}

\item{estimator_grid}{Lasso estimator grid matrix.}

\item{nbr_bootstraps}{Number of bootstrap simulations. Defaults to 1000.}
}
\value{
quantilesestimated Returns the value of the estimated quantile of the Lasso's effective noise.
}
\description{
'quantileestimated' lets you calculate the estimated quantiles using the criterion quantiles.
}
\examples{
lasso <- lasso(Y = Y,
               X = X,
               out = 1,
               numlam = 100,
               maxIter = 1e10,
               opTol = 1e-10,
               standardize = 0)

#Remove equidistant lambda row
BETA <- lasso \%>\% head(-1)

criterion_q <- criterionquantiles(Y = Y,
                                  X = X,
                                  level = c(0.90, 0.95, 0.99),
                                  estimator_grid = BETA,
                                  nbr_bootstraps = 10000,
                                  w = 20,
                                  bs_type = "stationary_block")
quantile_e <- quantileestimated(Y = Y,
                                X = X,
                                level = 0.95,
                                estimator_grid,
                                criterionquantiles = criterion_q)


}
