% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantileoracle}
\alias{quantileoracle}
\title{Calculates the oracle quantile}
\usage{
quantileoracle(n, p, kappa, rho, level, sdev = 1, nbr_bootstraps = 1000L)
}
\arguments{
\item{n}{Number of observations to generate for the simulated data.}

\item{p}{Number of covariates to generate for the simulated data.}

\item{kappa}{Covariance between the covariates.}

\item{rho}{The coefficient in the AR(1)-process. Controls the dependence of observation at time t to the observation at time t-1.}

\item{level}{Vector of levels.}

\item{sdev}{The specified standard deviation for the noise.}

\item{nbr_bootstraps}{Number of bootstrap simulations. Defaults to 1000.}
}
\value{
quantileoracle Returns the value of the approximated quantile of the lasso's effective noise.
}
\description{
'quantileoracle' lets you calculate the oracle quantiles with simulated data.
}
\examples{

quantileoracle(n = 500,
               p = 1000,
               kappa = 0.25,
               rho = 0.95
               level = c(0.95, 0.99),
               sdev = 1)

}
