% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantileoracleconditional}
\alias{quantileoracleconditional}
\title{Calculates the oracle quantile}
\usage{
quantileoracleconditional(X, level, sdev = 1, nbr_bootstraps = 1000L)
}
\arguments{
\item{X}{Covariates matrix.}

\item{level}{Vector of levels.}

\item{sdev}{The specified standard deviation for the noise.}

\item{nbr_bootstraps}{Number of bootstrap simulations. Defaults to 1000.}
}
\value{
quantileoracleconditional Returns the value of the approximated quantile of the lasso's effective noise dependent on X.
}
\description{
'quantileoracleconditinal' lets you calculate the oracle quantiles dependent on X.
}
\examples{

quantileoracleconditional(X = X,
                          level = c(0.95, 0.99),
                          sdev = 1)

}
