% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stationary_block_bootstrap}
\alias{stationary_block_bootstrap}
\title{Stationary Block Bootstrap}
\usage{
stationary_block_bootstrap(data, B = 1000L, w = 10L)
}
\arguments{
\item{data}{Data to bootstrap}

\item{B}{Number of bootstrap samples. Defaults to 1000.}

\item{w}{Mean number of blocks. Defaults to 10.}
}
\value{
Returns matrix of n*(B*p) dimensions.
}
\description{
'stationary_block_bootstrap' lets you compute moving block bootstrap samples of your data.
}
\examples{
bs_data <- stationary_block_bootstrap(data)

}
