% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoboot-package.R
\docType{package}
\name{lassoboot-package}
\alias{lassoboot}
\alias{lassoboot-package}
\title{lassoboot: Parametric Bootstrap for Lasso Regression with Measurement Uncertainty}
\description{
Provides a tidy, declarative API for parametric bootstrap inference on lasso
regression in the presence of measurement uncertainty. The primary workflow is
\code{lb_spec()} -> \code{lb_bootstrap()} -> \code{tidy()} / \code{autoplot()}.
}
\author{
\strong{Maintainer}: DoctorBear-it \email{barrett310@gmail.com}

Authors:
\itemize{
  \item DoctorBear-it \email{barrett310@gmail.com}
}

}
\keyword{internal}
