\name{LassoBoot-package}
\alias{LassoBoot-package}
\alias{LassoBoot}
\docType{package}
\title{
\packageTitle{LassoBoot}
}
\description{
\packageDescription{LassoBoot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LassoBoot}
\packageIndices{LassoBoot}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LassoBoot}

Maintainer: \packageMaintainer{LassoBoot}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
