% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{lb_folds_blocked}
\alias{lb_folds_blocked}
\title{Block fold generator}
\usage{
lb_folds_blocked(block, k = 5)
}
\arguments{
\item{block}{Column name (string) defining contiguous blocks assigned as
complete units to folds.}

\item{k}{Number of folds. Default \code{5}.}
}
\value{
A fold-generator closure \verb{function(data) -> integer vector}.
}
\description{
Block fold generator
}
