% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{lb_folds_custom}
\alias{lb_folds_custom}
\title{Custom fold generator}
\usage{
lb_folds_custom(fn)
}
\arguments{
\item{fn}{A function with signature \verb{function(data) -> integer vector} of
fold IDs the same length as \code{nrow(data)}. The function is responsible for
any re-randomization across CV repetitions.}
}
\value{
A fold-generator closure (the input \code{fn}, validated and classed).
}
\description{
Custom fold generator
}
