% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lb_plot_interactions}
\alias{lb_plot_interactions}
\title{Plot interaction effects}
\usage{
lb_plot_interactions(x, order = 2L, ...)
}
\arguments{
\item{x}{An \code{lb_boot} object.}

\item{order}{Interaction order to display. Default \code{2L}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot interaction effects
}
