% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lb_plot_selection}
\alias{lb_plot_selection}
\title{Plot selection probabilities}
\usage{
lb_plot_selection(x, threshold = 0.5, ...)
}
\arguments{
\item{x}{An \code{lb_boot} object.}

\item{threshold}{Reference line at this selection probability. Default \code{0.5}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot selection probabilities
}
